Clean up and move stuff
