+++ grakn-graql/src/main/java/ai/grakn/graql/internal/pattern/property/PropertyExecutor.java
@@ -21,7 +21,7 @@ package ai.grakn.graql.internal.pattern.property;
 import ai.grakn.concept.Concept;
 import ai.grakn.graql.Var;
 import ai.grakn.graql.admin.VarProperty;
-import ai.grakn.graql.internal.query.QueryOperationExecutor;
+import ai.grakn.graql.internal.query.runner.QueryOperationExecutor;
 import com.google.auto.value.AutoValue;
 import com.google.common.collect.ImmutableSet;
