+++ grakn-graql/src/main/java/ai/grakn/graql/internal/query/runner/ConceptBuilder.java
@@ -16,7 +16,7 @@
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
  */
 
-package ai.grakn.graql.internal.query;
+package ai.grakn.graql.internal.query.runner;
 
 import ai.grakn.concept.Attribute;
 import ai.grakn.concept.AttributeType;
